requirements.txt
faster-whisper>=1.0.0
openai>=1.14.0
PyQt6>=6.6.0
sounddevice>=0.4.6
numpy>=1.26.0
webrtcvad>=2.0.10
python-dotenv>=1.0.1
google-generativeai>=0.4.0